refactor(common): organize shared market data policies
