+++ src/libs/portfolio-common/portfolio_common/domain/market_data/fx_rate.py
@@ -1,3 +1,5 @@
+"""Foreign-exchange rate validation policy shared by calculation boundaries."""
+
 from decimal import Decimal
 
 from portfolio_common.domain.decimal_amount import decimal_or_none
